// massive-sprites/src/generated/wasm-loader.ts
export type WasmSource =
  | string
  | URL
  | ArrayBuffer
  | ArrayBufferView
  | Response;
// ...
export interface WasmInstance<T extends object> {
  readonly exports: T;
  readonly memory: WebAssembly.Memory;
  readonly table: WebAssembly.Table | undefined;
  readonly rawExports: WebAssembly.Exports;
  readonly module: WebAssembly.Module;
  readonly instance: WebAssembly.Instance;
  readonly initialize: (() => unknown) | undefined;
  readonly start: (() => unknown) | undefined;
}